pyproject.toml
[project]
name = "unite"
description = "UNified Integration Turbo Engine for spectroscopy"
authors = [{name = "Raphael Erik Hviding", email = "raphael.hviding@gmail.com"}]
requires-python = ">= 3.12"
version = "1.0.0"
dependencies = [
  "numpyro>=0.20.0,<0.21",
  "astropy>=7.2.0,<8",
  "pyyaml>=6.0.3,<7",
  ]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Astronomy",
  "Topic :: Scientific/Engineering :: Physics",
  "License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS :: MacOS X",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14"
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-cov>=6.0",
]
docs = [
  "sphinx>=8.0",
  "sphinx-autodoc-typehints>=2.0",
  "myst-parser>=4.0",
  "furo",
  "numpydoc",
]
dev = [
  "unite[test,docs]",
  "ruff>=0.11.9",
  "pre-commit",
]


[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["unite"]

# ---------------------------------------------------------------------------
# pixi
# ---------------------------------------------------------------------------

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64", "linux-64"]

[tool.pixi.environments]
default = {features = [], solve-group = "default"}
build = {features = ["build"], solve-group = "build"}
test = {features = ["test"], solve-group = "default"}
docs = {features = ["docs"], solve-group = "default"}

[tool.pixi.tasks]
format = "ruff format unite tests"
lint = "ruff check unite tests"
lint-fix = "ruff check --fix unite tests"

[tool.pixi.dependencies]
python = ">=3.13.3,<3.15"
ruff = ">=0.15.2,<0.16"

[tool.pixi.activation]
env = {"JAX_ENABLE_X64" = "1", "JAX_LOGGING_LEVEL" = "ERROR"}

[tool.pixi.pypi-dependencies]
unite = { path = ".", editable = true }

# -- build feature --
[tool.pixi.feature.build]
dependencies = { twine = ">=6.2.0,<7" }
pypi-dependencies = { build = ">=1.4.0, <2" }
tasks = { build = "python -m build" }

# -- test feature --
[tool.pixi.feature.test.pypi-dependencies]
pytest = ">=9.0.2, <10"
pytest-cov = ">=7.0.0, <8"

[tool.pixi.feature.test.tasks]
test = "pytest"
test-cov = "pytest --cov=unite --cov-report=term-missing --cov-report=xml:coverage.xml"

# -- docs feature --
[tool.pixi.feature.docs.pypi-dependencies]
sphinx = ">=9.1.0, <10"
sphinx-autodoc-typehints = ">=3.7.0, <4"
myst-parser = ">=5.0.0, <6"
furo = ">=2025.12.19, <2026"
numpydoc = ">=1.10.0, <2"

[tool.pixi.feature.docs.tasks]
docs = "sphinx-build -b html docs docs/_build/html"
docs-clean = "rm -rf docs/_build && sphinx-build -b html docs docs/_build/html"

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py312"
line-length = 88
extend-exclude = ["docs/conf.py"]

[tool.ruff.format]
quote-style = "single"
skip-magic-trailing-comma = true

[tool.ruff.lint]
extend-select = [
  "I",    # isort
  "N",    # pep8-naming
  "UP",   # pyupgrade
  "B",    # flake8-bugbear
  "SIM",  # flake8-simplify
  "RUF",  # ruff-specific rules
  "D",    # pydocstyle
  "NPY",  # numpy-specific rules
]
ignore = [
  "D100",  # missing docstring in public module
  "D102",  # missing docstring in public method (e.g. __new__)
  "D104",  # missing docstring in public package
  "D105",  # missing docstring in magic method
  "D203",  # conflicts with D211
  "D213",  # conflicts with D212
  "N802",  # function name should be lowercase (R is physics convention)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]

[tool.ruff.lint.isort]
known-first-party = ["unite"]
length-sort = false
combine-as-imports = true
split-on-trailing-comma = false
force-sort-within-sections = false

[tool.ruff.lint.pydocstyle]
convention = "numpy"

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-ra", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = ["error"]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------

[tool.coverage.run]
source = ["unite"]
branch = true

[tool.coverage.report]
show_missing = true
precision = 2
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "@abstractmethod",
  "raise NotImplementedError",
  "\\.\\.\\.",
]
